+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGenerator.kt
@@ -4,7 +4,6 @@ import com.wutsi.codegen.Context
 import com.wutsi.codegen.core.generator.CodeGenerator
 import com.wutsi.codegen.editconfig.EditorConfigCodeGenerator
 import com.wutsi.codegen.github.GitIgnoreCodeGenerator
-import com.wutsi.codegen.github.GithubWorkflowCodeGenerator
 import com.wutsi.codegen.kotlin.KotlinMapper
 import com.wutsi.codegen.kotlin.sdk.SdkModelCodeGenerator
 import io.swagger.v3.oas.models.OpenAPI
@@ -21,7 +20,7 @@ class ServerCodeGenerator(private val mapper: KotlinMapper) : CodeGenerator {
         ServerHerokuCodeGenerator(mapper).generate(openAPI, context)
 
         EditorConfigCodeGenerator().generate(openAPI, context)
-        GithubWorkflowCodeGenerator().generate(openAPI, context)
+        ServerGithubWorkflowCodeGenerator().generate(openAPI, context)
         GitIgnoreCodeGenerator().generate(openAPI, context)
     }
 }