+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerGithubWorkflowCodeGenerator.kt
@@ -1,11 +1,12 @@
-package com.wutsi.codegen.github
+package com.wutsi.codegen.kotlin.server
 
 import com.wutsi.codegen.Context
 import com.wutsi.codegen.core.generator.AbstractMustacheCodeGenerator
+import com.wutsi.codegen.kotlin.KotlinMapper
 import io.swagger.v3.oas.models.OpenAPI
 import java.io.File
 
-class GithubWorkflowCodeGenerator : AbstractMustacheCodeGenerator() {
+class ServerGithubWorkflowCodeGenerator : AbstractMustacheCodeGenerator() {
     override fun generate(openAPI: OpenAPI, context: Context) {
         generate("master.yml", openAPI, context)
         generate("pull_request.yml", openAPI, context)
@@ -13,6 +14,8 @@ class GithubWorkflowCodeGenerator : AbstractMustacheCodeGenerator() {
 
     override fun toMustacheScope(openAPI: OpenAPI, context: Context): Map<String, String?> =
         mapOf(
+            "artifactId" to ServerMavenCodeGenerator(KotlinMapper(context)).artifactId(context),
+            "version" to openAPI.info?.version,
             "jdkVersion" to context.jdkVersion,
             "secrets.GITHUB_TOKEN" to "{{secrets.GITHUB_TOKEN}}",
             "herokuApp" to context.herokuApp,