+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerHerokuCodeGenerator.kt
@@ -13,15 +13,27 @@ class ServerHerokuCodeGenerator(private val mapper: KotlinMapper) : AbstractMust
         "version" to openAPI.info?.version,
         "githubUser" to context.githubUser,
         "herokuApp" to context.herokuApp,
-        "description" to openAPI.info.version
+        "description" to openAPI.info.description
     )
 
     override fun generate(openAPI: OpenAPI, context: Context) {
         context.herokuApp ?: return
 
         generate(
-            inputPath = "/kotlin/server/Procfile.mustache",
+            inputPath = "/kotlin/server/heroku/Procfile.mustache",
             outputFile = File(context.outputDirectory, "Procfile"),
+            openAPI = openAPI,
+            context = context
+        )
+        generate(
+            inputPath = "/kotlin/server/heroku/app.json.mustache",
+            outputFile = File(context.outputDirectory, "app.json"),
+            openAPI = openAPI,
+            context = context
+        )
+        generate(
+            inputPath = "/kotlin/server/heroku/system.properties.mustache",
+            outputFile = File(context.outputDirectory, "system.properties"),
             openAPI = openAPI,
             context = context
         )