+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [0.0.11] 2021-03-17
 ### Added
-- Heroku deployment
+- Deployment of API server to Heroku
+### Changed
+- Server jar no longer deployed to github Maven repo
 
 ## [0.0.10] 2021-03-17
 ### Added